% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_utils.R
\name{get_clicked_row_value}
\alias{get_clicked_row_value}
\title{Get the value in the first column of the clicked row.}
\usage{
get_clicked_row_value()
}
\value{
HTML widget
}
\description{
Get the value in the first column of the clicked row.
}
\keyword{internal}
