% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly_run_session.R
\name{orderly_run_session}
\alias{orderly_run_session}
\title{Run an orderly report in separate session.}
\usage{
orderly_run_session(
  name = NULL,
  parameters = NULL,
  envir = NULL,
  root = NULL,
  locate = TRUE,
  message = NULL,
  instance = NULL,
  use_draft = FALSE,
  remote = NULL,
  tags = NULL,
  commit = FALSE
)
}
\arguments{
\item{name}{Name of the report to run (see
\code{\link[orderly]{orderly_list}}).  A leading \code{src/} will be
removed if provided, allowing easier use of autocomplete.
Alternatively, the default of \code{NULL} is useful if you have
already set the working directory to be the source directory.}

\item{parameters}{Parameters passed to the report. A named list of
parameters declared in the \code{orderly.yml}.  Each parameter
must be a scalar character, numeric, integer or logical.}

\item{envir}{The parent of the environment that will be used to
evaluate the report script; by default a new environment will be
made with the global environment as the parent.}

\item{root}{The path to an orderly root directory, or \code{NULL}
(the default) to search for one from the current working
directory if \code{locate} is \code{TRUE}.}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{message}{An optional character string containing a message
explaining why the report was run}

\item{instance}{Select instance of the source database to be used,
where multiple instances are configured.  Use a single
\emph{unnamed} character string to indicate an instance to
match.  If given, then this name must be present in all
databases where instances are listed in
\code{orderly_config.yml}, and will be ignored by all database
where instances are not given.  See the "orderly" vignette for
further information.}

\item{use_draft}{Should draft reports be used for dependencies?
This should be used only in development.  Valid values are
logical (\code{TRUE}, \code{FALSE}) or use the string
\code{newer} to use draft reports where they are newer than
archive reports.  For consistency, \code{always} and
\code{never} are equivalent to \code{TRUE} and \code{FALSE},
respectively.}

\item{remote}{Remote to use to resolve dependencies.  Use this in
order to run a report with the same dependencies as are
available on a remote server, particularly when using \code{id =
"latest"}.  Note that this is not the same as running
\code{\link[orderly]{orderly_pull_dependencies}}, then \code{orderly_run}
with \code{remote = NULL}, as the pull/run approach will use the
latest report in \emph{your} archive but the \code{remote =
"remote"} approach will use the latest approach in the
\emph{remote} archive (which might be less recent).}

\item{tags}{Character vector of tags to add to the report.  Tags
are immutable and cannot be removed once the report is run.
Tags added here will be \emph{in addition} to any tags listed in
the \code{tags:} field in \code{orderly.yml} and must be present
in \code{orderly_config.yml}.}

\item{commit}{If TRUE then report will be committed to archive after running}
}
\value{
Invisible return, called for side effects
}
\description{
This will run an orderly report using RStudio jobs. This will run the
report async i.e. it won't block your RStudio terminal. This means
that it returns no info about success of job run - it only tells us
the job has started.
}
