% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_addin.R
\name{start_addin}
\alias{start_addin}
\title{Addin for starting develop mode.}
\usage{
start_addin()
}
\description{
Allows users to see list of orderly reports and choose one to open in
development mode.
}
