% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_addin.R
\name{status_addin}
\alias{status_addin}
\title{Addin for viewing status of orderly develop mode.}
\usage{
status_addin()
}
\description{
Allows users to reload dependencies and to cleanup development mode.
}
